Current condition fragment layout changed

+++ app/src/main/java/com/breezee/sergeystasyuk/weatherinfo/pojos/dailyforecast/Imperial___.java
@@ -8,7 +8,7 @@ public class Imperial___ {
 
     @SerializedName("Value")
     @Expose
-    private Integer value;
+    private Float value;
     @SerializedName("Unit")
     @Expose
     private String unit;
@@ -16,11 +16,11 @@ public class Imperial___ {
     @Expose
     private Integer unitType;
 
-    public Integer getValue() {
+    public Float getValue() {
         return value;
     }
 
-    public void setValue(Integer value) {
+    public void setValue(Float value) {
         this.value = value;
     }
